MFH: - Use correct type (bug #32720)

+++ ext/snmp/snmp.c
@@ -20,7 +20,7 @@
   +----------------------------------------------------------------------+
 */
 
-/* $Id: snmp.c,v 1.70.2.18 2005/03/16 16:19:59 harrie Exp $ */
+/* $Id: snmp.c,v 1.70.2.19 2005/04/15 22:08:43 sniper Exp $ */
 
 #ifdef HAVE_CONFIG_H
 #include "config.h"
@@ -347,9 +347,9 @@ static void php_snmp_internal(INTERNAL_FUNCTION_PARAMETERS,
 	struct snmp_pdu *pdu=NULL, *response;
 	struct variable_list *vars;
 	oid name[MAX_NAME_LEN];
-	int name_length;
+	size_t name_length;
 	oid root[MAX_NAME_LEN];
-	int rootlen = 0;
+	size_t rootlen = 0;
 	int gotroot = 0;
 	int status, count;
 	char buf[2048];
